Added section screen and allowed for navigation to specific component
