+++ eml/components/courses/courseSubsription/StartNowButton.jsx
@@ -8,7 +8,7 @@ const StartNowButton = ({ course }) => {
 	const navigation = useNavigation();
 
 	const handlePress = () => {
-		navigation.navigate('Section', {
+		navigation.navigate('CourseOverview', {
 			course: course
 		});
 	};
